backend/requirements.txt
aiodns==3.0.0
aiohttp==3.8.1
aiohttp-socks==0.7.1
aiosignal==1.2.0
amqp==5.1.1
asgiref==3.5.2
asttokens==2.0.8
async-timeout==4.0.2
attrs==22.1.0
autopep8==1.7.0
backcall==0.2.0
billiard==3.6.4.0
celery==5.2.7
celery-redbeat==2.0.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudpickle==2.2.0
cryptography==37.0.4
dateparser==1.1.1
DateTime==4.7
debugpy==1.6.3
decorator==5.1.1
Deprecated==1.2.13
dill==0.3.5.1
dnspython==2.2.1
email-validator==1.2.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.0.0
Faker==14.1.1
flower==1.2.0
frozenlist==1.3.1
greenlet==1.1.3
humanize==4.3.0
idna==3.3
ipykernel==6.15.2
ipython==8.4.0
jedi==0.18.1
jsonschema==4.14.0
jupyter-core==4.11.1
jupyter_client==7.3.5
kombu==5.2.4
loguru==0.6.0
lxml==4.9.1
matplotlib-inline==0.1.6
mmh3==3.0.0
mpmath==1.2.1
multidict==6.0.2
nest-asyncio==1.5.5
numpy==1.23.2
openpyxl==3.0.10
packaging==21.3
pandas==1.4.3
pandas-ta==0.3.14b0
parso==0.8.3
passlib==1.7.4
pexpect==4.8.0
phonenumbers==8.12.54
pickleshare==0.7.5
pottery==3.0.0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
psutil==5.9.1
psycopg2-binary
ptyprocess==0.7.0
pure-eval==0.2.2
pwinput==1.0.2
pycares==4.2.2
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.9.2
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-socks==2.0.3
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
pyzmq==23.2.1
redis==4.3.4
regex==2022.3.2
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.40
SQLAlchemy-Utils==0.38.3
sqlparse==0.4.2
stack-data==0.5.0
stolgo==0.1.0
sympy==1.11
tenacity==8.0.1
toml==0.10.2
tornado==6.2
traitlets==5.3.0
typer==0.6.1
typing_extensions==4.3.0
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
websockets==10.3
wrapt==1.14.1
yarl==1.7.2
zope.interface==5.4.0
alembic
structlog==22.3.0
python-telegram-bot==20.3
gevent==22.10.2
ccxt==3.1.7
